Handle different Visual Studio variants better
Because Visual Studio likes to install to multiple different subfolders
depending upon which variant you install, build_detours.js should be
flexible enough to search all the normal variants be default, and allow
overriding if necessary.

Also, log a better error message if this process fails so that users
have an easier time fixing it themselves.